add call to checkLoggedIn
before actions that require it

+++ src/controller/ChildController.php
@@ -18,28 +18,32 @@ class ChildController extends Controller
 			}
 			return $this->view->render('home', [
 				'children' => $children
-				]);
+			]);
 		}
 	}
 
 	public function addChild(Parameter $post)
 	{
-		if ($post->get('submit')) {
-			$this->childDAO->addChild($post);
-			$this->session->set('add_child', 'Le nouvel enfant a bien été ajouté');
+		if ($this->checkLoggedIn()) {
+			if ($post->get('submit')) {
+				$this->childDAO->addChild($post);
+				$this->session->set('add_child', 'Le nouvel enfant a bien été ajouté');
+				header('Location: index.php?route=listChildren');
+				exit;
+			}
+			return $this->view->render('register');
+		}
+	}
+
+	public function deleteChild($childId)
+	{
+		if ($this->checkLoggedIn()) {
+			$this->childDAO->deleteChild($childId);
+			$this->session->set('delete_child', 'L\'enfant a bien été supprimé');
 			header('Location: index.php?route=listChildren');
 			exit;
 		}
-		return $this->view->render('register');
 	}
-	
-	public function deleteChild($childId)
-	{
-		$this->childDAO->deleteChild($childId);
-		$this->session->set('delete_child', 'L\'enfant a bien été supprimé');
-		header('Location: index.php?route=listChildren');
-		exit;
-	}	
 	/**
 	 * childCard 
 	 *
@@ -48,10 +52,12 @@ class ChildController extends Controller
 	 */
 	public function childCard($childId)
 	{
-		$childArray = $this->childDAO->childCard($childId);
-		$child = new Child($childArray);
-		return $this->view->render('card', [
-			'child' => $child
-		]);
+		if ($this->checkLoggedIn()) {
+			$childArray = $this->childDAO->childCard($childId);
+			$child = new Child($childArray);
+			return $this->view->render('card', [
+				'child' => $child
+			]);
+		}
 	}
 }
